+++ node/globals.ts
@@ -20,6 +20,7 @@ declare global {
     rootPath: string
     matchingBindings: Binding[]
     bindingAddress?: string
+    isCrossBorder: boolean
     nextEvent:  {
       event: string,
       payload: Events