+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+ 
+### Added
+
+- `MAX_CALL_STACK_SIZE` when appending entries to XML file
 
 ## [2.16.4] - 2024-09-11