+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Updated README.md file
+
 ## [2.16.6] - 2025-01-30
 
 ## [2.16.5] - 2025-01-23